requirements.txt
streamlit
pandas
numpy
plotly
requests
python-dateutil
ccxt
